+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/Context.java
@@ -999,8 +999,39 @@ public final class Context implements AutoCloseable {
          * @since 1.0
          */
         public Builder logHandler(final Handler logHandler) {
-            Objects.requireNonNull(logHandler, "Hanlder must be non null.");
+            Objects.requireNonNull(logHandler, "Handler must be non null.");
             this.customLogHandler = logHandler;
+            return this;
+        }
+
+        /**
+         * Installs a new logging {@link Handler} using given {@link OutputStream}. The logger's
+         * {@link Level} configuration is done using the {@link #options(java.util.Map) Context's
+         * options}. The level option key has the following format:
+         * {@code log.languageId.loggerName.level} or {@code log.instrumentId.loggerName.level}. The
+         * value is either the name of pre-defined {@link Level} constant or a numeric {@link Level}
+         * value. If not explicitly set in options the level is inherited from the parent logger.
+         * <p>
+         * <b>Examples</b> of setting log level options:<br>
+         * {@code builder.option("log.level","FINE");} sets the {@link Level#FINE FINE level} to all
+         * {@code TruffleLogger}s.<br>
+         * {@code builder.option("log.js.level","FINE");} sets the {@link Level#FINE FINE level} to
+         * JavaScript {@code TruffleLogger}s.<br>
+         * {@code builder.option("log.js.com.oracle.truffle.js.parser.JavaScriptLanguage.level","FINE");}
+         * sets the {@link Level#FINE FINE level} to {@code TruffleLogger} for the
+         * {@code JavaScriptLanguage} class.<br>
+         * <p>
+         * If the {@code logHandler} is not set on {@link Engine} nor on {@link Context} the log
+         * messages are printed to {@link #out(java.io.OutputStream) Context's standard output
+         * stream}.
+         *
+         * @param logOut the {@link OutputStream} to use for logging in built {@link Context}.
+         * @return the {@link Builder}
+         * @since 1.0
+         */
+        public Builder logHandler(final OutputStream logOut) {
+            Objects.requireNonNull(logOut, "LogOut must be non null.");
+            this.customLogHandler = new PolyglotStreamHandler(logOut, true, true);
             return this;
         }