+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/Engine.java
@@ -439,8 +439,35 @@ public final class Engine implements AutoCloseable {
          * @since 1.0
          */
         public Builder logHandler(final Handler logHandler) {
-            Objects.requireNonNull(logHandler, "Hanlder must be non null.");
+            Objects.requireNonNull(logHandler, "Handler must be non null.");
             this.customLogHandler = logHandler;
+            return this;
+        }
+
+        /**
+         * Installs a new logging {@link Handler} using given {@link OutputStream}. The logger's
+         * {@link Level} configuration is done using the {@link #options(java.util.Map) Engine's
+         * options}. The level option key has the following format:
+         * {@code log.languageId.loggerName.level} or {@code log.instrumentId.loggerName.level}. The
+         * value is either the name of pre-defined {@link Level} constant or a numeric {@link Level}
+         * value. If not explicitly set in options the level is inherited from the parent logger.
+         * <p>
+         * <b>Examples</b> of setting log level options:<br>
+         * {@code builder.option("log.level","FINE");} sets the {@link Level#FINE FINE level} to all
+         * {@code TruffleLogger}s.<br>
+         * {@code builder.option("log.js.level","FINE");} sets the {@link Level#FINE FINE level} to
+         * JavaScript {@code TruffleLogger}s.<br>
+         * {@code builder.option("log.js.com.oracle.truffle.js.parser.JavaScriptLanguage.level","FINE");}
+         * sets the {@link Level#FINE FINE level} to {@code TruffleLogger} for the
+         * {@code JavaScriptLanguage} class.<br>
+         *
+         * @param logOut the {@link OutputStream} to use for logging in engine's {@link Context}s.
+         * @return the {@link Builder}
+         * @since 1.0
+         */
+        public Builder logHandler(final OutputStream logOut) {
+            Objects.requireNonNull(logOut, "LogOut must be non null.");
+            this.customLogHandler = new PolyglotStreamHandler(logOut, true, true);
             return this;
         }
 
@@ -540,6 +567,11 @@ public final class Engine implements AutoCloseable {
         @Override
         public StackFrame newPolyglotStackTraceElement(PolyglotException e, AbstractStackFrameImpl impl) {
             return e.new StackFrame(impl);
+        }
+
+        @Override
+        public Handler newLogHandler(OutputStream logOut, boolean closeStream, boolean flushOnPublish) {
+            return new PolyglotStreamHandler(logOut, closeStream, flushOnPublish);
         }
     }