+++ sdk/CHANGELOG.md
@@ -8,6 +8,8 @@ This changelog summarizes major changes between Graal SDK versions. The main foc
 
 ## Version 1.0 RC7
 * Graal SDK was relicensed from GPLv2 with CPE to Universal Permissive License (UPL).
+## Version 1.0 RC7
+* Added a [Context.Builder.logHandler](http://www.graalvm.org/sdk/javadoc/org/graalvm/polyglot/Context.Builder.html#logHandler-java.io.OutputStream-) and [Engine.Builder.logHandler](http://www.graalvm.org/sdk/javadoc/org/graalvm/polyglot/Engine.Builder.html#logHandler-java.io.OutputStream-) methods to install a logging handler writing into a given `OutputStream`
 
 ## Version 1.0 RC6
 * Added new `ByteSequence` utility to the IO package that is intended to be used as immutable byte sequence representation.